wip: improved input source handles

+++ teleop/include/teleop/inputs/Axis.hpp
@@ -11,13 +11,13 @@
 namespace teleop
 {
 
-class Axis : public InputCommon<double>
+class Axis : public InputCommon<float>
 {
 public:
   using SharedPtr = std::shared_ptr<Axis>;
   using WeakPtr = std::weak_ptr<Axis>;
 
-  explicit Axis(std::string name) : InputCommon<double>(std::move(name))
+  explicit Axis(std::string name) : InputCommon<float>(std::move(name))
   {
   }
 };
